Adds example variant file.
This file is an example of what the final file of called variants
should look like in order to be graphed for bottleneck and
de novo variant analysis. Note the final column is for now
added by hand in excel by cross-checking the called variants with
primer binding sites and excluding variants that fall within that
region.